Added code to generate amf file
It successfully reads .blox into a string list, converts it to a string.
I can find frame that will be converted, it's associated dimensions,
store in hashmap, find next frame associated with the frame about to be
printed, and I need to add in some logic to find dimensions if frame is
assigned with a frame and not explicit dimensions. I can print header,
iterate through hashmap, and print footer to amf file.

It's incomplete and not very efficient at the moment.

Also added prior iterations of the code